Maintenance Release with U-series work-in-progress. (#146)
* updates copyright notice

* updates McLean Regression version

* turned off McLeanRegression file outputs

* updates pom to snapshot versions of topsoil and mcleanregression

* updates pom

* fixes show any 2 bug found by Noah

* updates xstream reference

* refactors note button fro fractions

* adds bifurcation of useries into carbonate and igneous

* advances U-series igneous

* begins visualizations of useries igneous

* adds any2 to uth sortof

* refactors specialty plots

* adds sample date model for Th isochron

* refines behavior of useries isochrons

* advances isochron details

* adds gui for selecting isochrons

* refines isochron handling

* refactors

* customizes behavior of sample dates maanger based on sample type

* fixes filtering bug for Useries

* improves sample dates manager

* improves sample dates manager

* updates physical constants to include 232Ra and uses in UTh calcs.

* fixes scroll bug and refactors

+++ src/main/java/org/earthtime/UPb_Redux/dateInterpretation/WeightedMeanGraphPanel.java
@@ -154,12 +154,12 @@ public class WeightedMeanGraphPanel extends JPanel
 
     private double mapX(double x, double minX, double rangeX, int widthG) {
 
-        return (((x - minX) / rangeX) * (double) widthG) + (double) getLeftMargin();
+        return (((x - minX) / rangeX) * widthG) + getLeftMargin();
     }
 
     private double mapY(double y, double maxY, double rangeY, int heightG) {
 
-        return (((maxY - y) / rangeY) * (double) heightG) + (double) getTopMargin();
+        return (((maxY - y) / rangeY) * heightG) + getTopMargin();
     }
 
     /**
